requirements/dev.txt
ipython
coverage
ipdb
Fabric3
